PRO-413: New field user education Added new table: `UserEducation` (M:O) Custom migrations for `organization` -> `UserEducation`: ``` python manage.py migrate_education python manage.py reverse_migrate_education ``` The migrations themselves are "safe" because they affect only the new entity, without erasing or editing existing information.
`organization` - the built field for education functionality is fully preserved, the front must be ready for the new format.
Accordingly, to carry out the migration, the front must be ready for the new format of filling in information on education.
After all migrations and when the front is ready, it (the `organization` field) will simply need to be deleted.